+++ shared/ui/src/commonMain/kotlin/com/shub39/rush/shared/ui/viewmodels/LyricsVM.kt
@@ -33,6 +33,7 @@ import com.shub39.rush.shared.ui.lyrics.breakLyrics
 import com.shub39.rush.shared.ui.lyrics.toSongUi
 import com.shub39.rush.shared.ui.sortMapByKeys
 import kotlin.time.Duration.Companion.milliseconds
+import kotlin.time.TimeSource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
@@ -464,10 +465,10 @@ class LyricsVM(
             viewModelScope.launch(Dispatchers.Default) {
                 combine(MediaListener.songPositionFlow, MediaListener.playbackSpeedFlow, ::Pair)
                     .collectLatest { (position, speed) ->
-                        val start = System.currentTimeMillis()
+                        val start = TimeSource.Monotonic.markNow()
 
                         while (isActive) {
-                            val elapsed = (speed * (System.currentTimeMillis() - start)).toLong()
+                            val elapsed = (speed * start.elapsedNow().inWholeMilliseconds).toLong()
 
                             _playbackInfo.update {
                                 it.copy(